Add files via upload
vadd, vdiv, vmult, vpol, and vcart are all generic functions for handling operations with vectors in both a polar and cartesian format. They are written to make circuit analysis easier, but can be used for other things as well. paraResist determines the resistance of a parallel circuit based on the input series elements.